changed the package name of the app and added application-wide exception handler
