+++ app/src/main/java/com/superiorapp/SuperStore/App.java
@@ -1,0 +1,96 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package com.superiorapp.SuperStore;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import java.io.FileReader;
+
+import com.opencsv.CSVReader;
+import com.opencsv.CSVReaderBuilder;
+import com.opencsv.bean.CsvToBeanBuilder;
+
+import com.superiorapp.SuperStore.entities.Row;
+import com.superiorapp.SuperStore.ui.UI;
+
+import javafx.application.Platform;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
+
+public final class App {
+
+    /**
+     * Private constructor.
+     */
+    private App() {
+
+    }
+
+    /**
+     * main class of the application.
+     * 
+     * @param args args.
+     */
+    public static void main(final String[] args) throws Exception {
+        if (args.length == 0) {
+            System.out.println("Error: No arguments provided.");
+            System.out.println(
+                    "For instructions on how to run the application please check https://github.com/RealLuckyDeveloper/SuperStoreOrders-Java/blob/main/README.md");
+            System.exit(1);
+        }
+        List<Row> beans = null;
+        List<String> returnIDs = null;
+        try {
+            beans = new CsvToBeanBuilder<Row>(new FileReader(args[0]))
+                    .withSeparator(';')
+                    .withType(Row.class)
+                    .build()
+                    .parse();
+
+            CSVReader reader = new CSVReaderBuilder(new FileReader(args[1])).build();
+            returnIDs = reader.readAll()
+                    .stream()
+                    .skip(1) // discard column titles and remove "Yes;" at the start
+                    .map(array -> array[0].substring(4))
+                    .collect(Collectors.toList());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        if (beans == null) {
+            throw new Exception("main file data from csv is null!!!");
+        }
+        if (returnIDs == null) {
+            throw new Exception("return file data from csv is null!!!");
+        }
+
+        // set up exception handler for application
+        Thread.setDefaultUncaughtExceptionHandler((thread, cause) -> {
+            try {
+                cause.printStackTrace();
+
+                String errorMessage = cause.getMessage();
+
+                final Runnable showDialog = () -> {
+                    Alert dialog = new Alert(AlertType.ERROR, errorMessage, ButtonType.OK);
+                    dialog.showAndWait();
+                };
+
+                if (Platform.isFxApplicationThread()) {
+                    showDialog.run();
+                } else {
+                    Platform.runLater(showDialog);
+                }
+            } catch (Throwable t) {
+                // Handle any exceptions that occur during dialog creation
+            } finally {
+                System.exit(-1); // Exit the application
+            }
+        });
+
+        UI.passDataAndLaunch(beans, returnIDs);
+    }
+}
