+++ app/src/main/java/com/superiorapp/SuperStore/ui/util/MapEntry.java
@@ -1,4 +1,4 @@
-package org.example.ui.util;
+package com.superiorapp.SuperStore.ui.util;
 
 public class MapEntry {
     private Object key;
