+++ app/src/main/java/com/superiorapp/SuperStore/util/DoubleConverter.java
@@ -1,4 +1,4 @@
-package org.example.util;
+package com.superiorapp.SuperStore.util;
 
 import com.opencsv.bean.AbstractBeanField;
 import com.opencsv.exceptions.CsvDataTypeMismatchException;
