+++ app/src/test/java/com/superiorapp/SuperStore/AppTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package org.example;
+package com.superiorapp.SuperStore;
 
 // import org.junit.jupiter.api.Test;
 // import static org.junit.jupiter.api.Assertions.*;
