+++ settings.gradle.kts
@@ -10,5 +10,5 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.7.0"
 }
 
-rootProject.name = "final-exercise"
+rootProject.name = "SuperStore"
 include("app")
